# frontend/openapi/citizenvoice/openapi.yaml
openapi: 3.0.3
info:
  title: CitizenVoice API
  version: 2.0.1
  description: Documentation of API endpoints for CitizenVoice
paths:
...
  /api/v2/questions/:
    get:
      operationId: v2_questions_list
      description: |-
        Question ViewSet used to query data from database.
        The `create` method is overwritten to accept one data object or an array of objects.
      tags:
      - v2
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
          description: ''
    post:
      operationId: v2_questions_create
      description: Here we are overwriting the default create method from the Django
        REST framework to update or create Questions by list or by single instances
      tags:
      - v2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Question'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Question'
        required: true
      security:
      - tokenAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
          description: ''
...
  /api/v2/questions/{id}/ordered_questions/:
    get:
      operationId: v2_questions_ordered_questions_retrieve
      description: |-
        Retrieve a list of questions for a given survey, ordered by the 'order' field.
        API url: `/api/questions/{survey_id}/ordered_questions`

        Parameters:
            request (Request): The request object used to make the API call.
            pk = survey_id (int): The primary key of the Survey instance to retrieve questions for.

        Returns:
            Response: A JSON response containing a list of serialized Question instances.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this question.
        required: true
      tags:
      - v2
      security:
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
          description: ''
...
components:
  schemas:
...
    Question:
      type: object
      description: |-
        Serializes 'text', 'order', 'required', 'question_type', 'choices', 'is_geospatial', 'map_view'
        fields of the Question model for the API.
      properties:
        id:
          type: integer
          readOnly: true
        url:
          type: string
          format: uri
          readOnly: true
        text:
          type: string
          title: Text of the Question
        order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          title: Order of where question is placed
        required:
          type: boolean
          default: true
          title: Question must be filled out
        question_type:
          allOf:
          - $ref: '#/components/schemas/QuestionTypeEnum'
          default: text
          title: Type of question
        choices:
          type: string
          nullable: true
          title: Choices for answers
        survey:
          type: string
          format: uri
          readOnly: true
        is_geospatial:
          type: boolean
          default: false
          title: If the question must be answered geospatially or not
        mapview:
          type: string
          format: uri
          nullable: true
      required:
      - id
      - order
      - survey
      - text
      - url
    QuestionTypeEnum:
      enum:
      - text
      - short-text
      - radio
      - select
      - select-multiple
      - select_image
      - integer
      - float
      - date
      type: string
      description: |-
        * `text` - text (multiple line)
        * `short-text` - short text (one line)
        * `radio` - radio
        * `select` - select
        * `select-multiple` - Select Multiple
        * `select_image` - Select Image
        * `integer` - integer
        * `float` - float
        * `date` - date
...
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"